18. 리뷰 리스트 구현

+++ app/src/main/java/aop/fastcampus/part6/chapter01/model/restaurant/RestaurantReviewModel.kt
@@ -9,6 +9,6 @@ data class RestaurantReviewModel(
     override val type: CellType = CellType.REVIEW_CELL,
     val title: String,
     val description: String,
-    val grade: Int,
+    val grade: Float,
     val thumbnailImageUri: Uri? = null
 ): Model(id, type)
