+++ app/src/main/java/aop/fastcampus/part6/chapter01/widget/adapter/viewholder/review/RestaurantReviewViewHolder.kt
@@ -34,7 +34,7 @@ class RestaurantReviewViewHolder(
 
             reviewTitleText.text = model.title
             reviewText.text = model.description
-            ratingBar.rating = model.grade.toFloat()
+            ratingBar.rating = model.grade
         }
     }
